ReadMe
In this program, posting lists is generated from the RCV1 news corpus
(http://www.daviddlewis.com/resources/testcollections/rcv1/). Index is
rebuilt after reading in the data file term.idx. Linked List is used to
store the index data in memory. Two indexes are constructed with two
different ordering strategies: with one strategy, the posting of each
term should be ordered by increasing document IDs; with the other
strategy, the postings of each term should be ordered by decreasing term
frequencies. After that, modules are implemented that return documents
based on term-at-a-time with the postings list ordered by term
frequencies,
and document-at-a-time with the postings list ordered by doc IDs for a
set of queries. Java is used for this assignment.